Improve performance significantly
This change transfers the image to the PowerShell server as binary data
rather than a base64-encoded data URL. This means there are fewer bytes
to transfer.

More significantly, we now pass the byte to write to the file to
Out-File's -Value parameter rather than piping the bytes in one at a
time.